+++ src/main/java/pl/pwr/ite/dynak/DataProcessingUtils/TipperLorry.java
@@ -16,7 +16,7 @@ public class TipperLorry {
             String data;
             while ((data = reader.readLine()) != null) {
                 String[] numbers = data.split(", ");
-                Bucket bucket = new Bucket(Integer.parseInt(numbers[0]), Integer.parseInt(numbers[1]), Integer.parseInt(numbers[2]));
+                Bucket bucket = new Bucket(Integer.parseInt(numbers[0]), Double.parseDouble(numbers[1]), Integer.parseInt(numbers[2]));
                 buckets.add(bucket);
             }
             return buckets;
