+++ src/main/java/pl/pwr/ite/dynak/DataProcessingUtils/TowerSpot.java
@@ -8,9 +8,9 @@ import lombok.Setter;
 public class TowerSpot {
     private Integer spotNumber;
     private Integer radius;
-    private HashMap<Double, Integer> points = new HashMap<>();
+    private HashMap<Integer, Double> points = new HashMap<>();
     private Double totalPoints;
-    //HashMap takes the points as the key and the bucket numbers as their values
+    //HashMap takes the bucket numbers as their keys and the points as the values
     public TowerSpot(Integer spotNumber, Integer radius) {
         this.points = new HashMap<>();
         this.totalPoints = 0.0;
